vision/requirements.txt
Flask==2.0.2
google-api-python-client==1.12.8
google-cloud-core==2.1.0
google-cloud-storage==1.42.3
google-cloud-vision==2.6.0
numpy==1.21.2
# opencv-python==4.5.4.58
gunicorn==20.1.0
opencv-python-headless==4.5.4.58